+++ apis/python/src/tiledbsoma/_geometry_dataframe.py
@@ -42,6 +42,7 @@ from ._spatial_util import (
     process_spatial_df_region,
 )
 from ._types import OpenTimestamp
+from ._util import _cast_record_batch
 from .options import SOMATileDBContext
 from .options._util import build_clib_platform_config
 
@@ -566,7 +567,10 @@ class GeometryDataFrame(SpatialDataFrame, somacore.GeometryDataFrame):
         outline_transformer = clib.OutlineTransformer(coordinate_space_to_json(self._coord_space))
         for batch in batches:
             table = (
-                clib.TransformerPipeline(batch.cast(batch_schema, safe=True)).transform(outline_transformer).asTable()
+                clib
+                .TransformerPipeline(_cast_record_batch(batch, batch_schema, safe=True))
+                .transform(outline_transformer)
+                .asTable()
             )
             for subbatch in table.to_batches():
                 mq = ManagedQuery(self)._handle